Changed long* to int* for compatibility on unix (R int is 4 byte, long on unix is 8)

+++ src/interface.cpp
@@ -52,8 +52,9 @@ extern "C"
 	//             param4 = number of bootstrap samples
 	//  dbg - if non-zero, show debugging information
    
-	//XXXvoid r_oja(long* rows,long* cols,double* data,double* vec_out,double* mat_out,long* func,double* param1, double* param2, long* param3, long* param4,long* dbg, long* rSeed)
-	void r_oja(long* rows,long* cols,double* data,double* vec_out,double* mat_out,long* func,double* param1, double* param2, long* param3, long* param4,long* dbg)
+	//XXXvoid r_oja(int* rows,int* cols,double* data,double* vec_out,double* mat_out,int* func,double* param1, double* param2, int* param3, int* param4,int* dbg, int* rSeed)
+	// change: switched from using long type to int, could cause errors in unix, as input was 4 byte and long is 8 byte in unix
+	void r_oja(int* rows,int* cols,double* data,double* vec_out,double* mat_out,int* func,double* param1, double* param2, int* param3, int* param4,int* dbg)
 	{
 		int dim=(int)*cols;
 		int size=(int)*rows;
